+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Dependencies
 
+- Bump `dangoslen/dependabot-changelog-helper@releases/v4.1` from 4.1 to 4.2 ([#408](https://github.com/dangoslen/dependabot-changelog-helper/pull/408))
 - Bump `typescript` from 5.8.3 to 5.9.2 ([#411](https://github.com/dangoslen/dependabot-changelog-helper/pull/411))
 
 ### Fixed